simplifying situation analysis.
