+++ rj_gameplay/rj_gameplay/gameplay_node.py
@@ -8,30 +8,32 @@ and allows the PlaySelector to be changed between Test and other forms.
 
 from typing import List, Optional, Tuple
 
+import numpy as np
 import rclpy
-from rclpy.node import Node
-from rclpy.qos import QoSProfile, QoSDurabilityPolicy
-
-from rj_msgs import msg
-from rj_geometry_msgs import msg as geo_msg
-from std_msgs.msg import String as StringMsg
-
+import stp.local_parameters as local_parameters
+import stp.play
 import stp.rc as rc
-import stp.utils.world_state_converter as conv
 import stp.situation as situation
 import stp.skill
-import stp.play
+import stp.utils.world_state_converter as conv
+from rclpy.node import Node
+from rclpy.qos import QoSDurabilityPolicy, QoSProfile
+from rj_geometry_msgs import msg as geo_msg
+from rj_msgs import msg
+from std_msgs.msg import String as StringMsg
 from stp.action import IAction
-
-import stp.local_parameters as local_parameters
 from stp.global_parameters import GlobalParameterClient
 
-import numpy as np
+import rj_gameplay.basic_play_selector as basic_play_selector
 from rj_gameplay.action.move import Move
 
 # ignore "unused import" error
-from rj_gameplay.play import line_up, basic_defense, keepaway, basic122  # noqa: F401
-import rj_gameplay.basic_play_selector as basic_play_selector
+from rj_gameplay.play import (  # noqa: F401
+    basic_defense,
+    basic_offense,
+    keepaway,
+    line_up,
+)
 
 NUM_ROBOTS = 16
 
@@ -53,7 +55,8 @@ class GameplayNode(Node):
         super().__init__("gameplay_node")
 
         # do not change this line, change the test play passed in at bottom of file
-        self.test_play = test_play
+        self._curr_play = test_play
+        self._curr_situation = None
 
         self.world_state_sub = self.create_subscription(
             msg.WorldState,
@@ -213,15 +216,22 @@ class GameplayNode(Node):
             assert self.world_state is not None
 
     def gameplay_tick(self) -> None:
+        """
+        Get situation, play from self.play_selector and update the currently running play if needed.
+        Then, add field and game_info to world_state, and push global obstacles to motion planning.
+        """
         self.update_world_state()
 
         if self.world_state is not None:
-            if self.test_play is None:
-                curr_situation, curr_play = self.play_selector.select(self.world_state)
-                intents = curr_play.tick(self.world_state)
-            else:
-                intents = self.test_play.tick(self.world_state)
-                curr_play = self.test_play
+            new_situation, new_play = self.play_selector.select(self.world_state)
+
+            if type(self._curr_play) != type(new_play) or type(
+                self._curr_situation
+            ) != type(new_situation):
+                self._curr_play = new_play
+                self._curr_situation = new_situation
+
+            intents = self._curr_play.tick(self.world_state)
 
             if intents:
                 for i in range(len(self.world_state.our_robots)):
@@ -241,7 +251,7 @@ class GameplayNode(Node):
             self.add_ball_to_global_obs(global_obstacles, game_info)
 
             self.global_obstacles_pub.publish(global_obstacles)
-            self.debug_callback(curr_play, curr_play.prioritized_tactics)
+            self.debug_callback(self._curr_play, self._curr_play.prioritized_tactics)
         else:
             self.get_logger().warn("World state was none!")
 
@@ -461,7 +471,7 @@ def main():
 
     # change this line to test different plays (set to None if no desired test play)
 
-    test_play = basic122.Basic122()
+    test_play = None
 
     gameplay = GameplayNode(play_selector, test_play)
     rclpy.spin(gameplay)
